Made buildTree() and processArray() functions
Finished buildTree() and processArray() function
logic. processArray() is a separate function
that takes in an array and sorts it and removes
any duplicate entries before passing that array
to the buildTree function which constructs the
left and right subtrees before returning the middle
root node to the processArray() function and
assigns the resulting root node to the root variable.