Finished insert() function
It uses recursion to insert a new leaf node
according to the binary tree structure. The
basecase is when we reach an empty node
(a leaf node) and we keep traversing the tree until
we reach that empty node, which will then return the
input node to be assigned to wherever we are at,
otherwise returns the same node to keep track of the
original node history.