Finished deletion() function
The deletion function is confusing but it uses the
same idea as the insertion, but with more
operations when it hits the basecase, which is
when currentRoot equals the node we are looking to
delete. When we reach the node to delete, we
have 3 scenarios depending on the deleted nodes
children: if it has no children, we can assume
that it's a leaf node so no other changes are
needed except making the node itself null. If it
has at least one child (left or right, doesn't
matter) then we need to make sure to replace
the currentRoot with the child (move the child
up). If the deleted item has BOTH children,
you have to replace it with the
inorder successor, which is the next higher value
than the node we are deleting so it should fit
the BST structure. (essentailly the inorder successor)
is the next higher value after the node we selected
to delete, so we need to traverse the tree
to get that value to replace it where the
deleted node is at. Also the reason we traverse the
tree in the way of right subtree then left
as much as possible to get the inorder successor
is because since we are finding the next
number that comes after the one we are deleting,
we know that all numbers in the right subtree are
greater than the node we are deleting, so
we travel right once, then travel left because
remember, if we go right it would mean we would have
values greater than the one we are looking for, we
are just looking for the next higher value after the
current node we are deleting, so traveling leftmost
will always yield that value since the smallest
value on the right subtree will be on the very
left side (or keep going left until we can't)