Finished the isBalanced() function
isBalanced uses postOrder DFS to count the
height of every node and on every node
check the left and right heights and if
ANY of those left and right heights differences
are greater than 1 (after absolute value), then
we can automatically flag this tree as
not balanced. Pretty much traverse the tree
at every node and compare the entire left
side and the entire right side's height together.